升级包

+++ src/Masuit.MyBlogs.Core/Models/ViewModel/PostModelBase.cs
@@ -2,19 +2,18 @@
 
 public class PostModelBase
 {
-	/// <summary>
-	/// id
-	/// </summary>
-	public int Id { get; set; }
+    /// <summary>
+    /// id
+    /// </summary>
+    public int Id { get; set; }
 
-	/// <summary>
-	/// 标题
-	/// </summary>
-	public string Title { get; set; }
+    /// <summary>
+    /// 标题
+    /// </summary>
+    public string Title { get; set; }
 
-	/// <summary>
-	/// 浏览次数
-	/// </summary>
-	public int ViewCount { get; set; }
-
+    /// <summary>
+    /// 浏览次数
+    /// </summary>
+    public long ViewCount { get; set; }
 }
